+++ karma.conf.js
@@ -16,6 +16,7 @@ module.exports = (config) => {
     },
     files: [  
             'src/spec/_boot.ts', 
+            { pattern: 'src/spec/images/**/*.mp3', included: false, served: true },
             { pattern: 'src/spec/images/**/*.png', included: false, served: true },
             { pattern: 'src/spec/images/**/*.gif', included: false, served: true },
             { pattern: 'src/spec/images/**/*.txt', included: false, served: true },
@@ -50,7 +51,7 @@ module.exports = (config) => {
             use: ['css-loader']
           },
           {
-            test: /\.(png|jpg|gif)$/i,
+            test: /\.(png|jpg|gif|mp3)$/i,
             use: [
               {
                 loader: 'url-loader',
@@ -101,7 +102,7 @@ module.exports = (config) => {
     customLaunchers: {
       ChromeHeadless_with_audio: {
           base: 'ChromeHeadless',
-          flags: ['--autoplay-policy=no-user-gesture-required']
+          flags: ['--autoplay-policy=no-user-gesture-required', '--mute-audio']
       },
       ChromeHeadless_with_debug: {
         base: 'ChromeHeadless',