+++ src/spec/TileMapSpec.ts
@@ -64,7 +64,7 @@ describe('A TileMap', () => {
       tm.registerSpriteSheet('default', spriteTiles);
       tm.data.forEach(function (cell: ex.Cell) {
         cell.solid = true;
-        cell.pushSprite(new ex.TileSprite('default', 0));
+        cell.pushSprite(spriteTiles.sprites[0]);
       });
       tm._initialize(engine);
 
@@ -91,7 +91,7 @@ describe('A TileMap', () => {
       tm.registerSpriteSheet('default', spriteTiles);
       tm.data.forEach(function (cell: ex.Cell) {
         cell.solid = true;
-        cell.pushSprite(new ex.TileSprite('default', 0));
+        cell.pushSprite(spriteTiles.sprites[0]);
       });
       tm._initialize(engine);